implemented Collaboration mode

+++ src/sensing/onAgentCallback.js
@@ -8,10 +8,9 @@ const AGENT_PROBABILITY_THRESHOLD = 0.7;
 /**
  * @param {AgentData[]} agents
  * @param {BeliefSet} beliefs
- * @param {DeliverooApi} client
  * @returns {Promise<void>}
  */
-export async function onAgentCallback_simple(agents,beliefs, client) {
+export async function onAgentCallback_simple(agents,beliefs) {
 	//console.log("called onAgentCallback_simple");
 	for(let existing_belief of beliefs.agentBeliefs){
 		let new_data = agents.find((a) => a.id === existing_belief.id);
@@ -25,15 +24,11 @@ export async function onAgentCallback_simple(agents,beliefs, client) {
 
 			agents.splice(agents.indexOf(new_data),1);
 		} else {
-			// infer agent belief
-
-			// should consider the believed intention and execute a step of it
-			// in this implementation we will just reduce the probability of an agent being there
-			
 			existing_belief.probability = existing_belief.probability - AGENT_PROBABILITY_DECAY;
 			existing_belief.time = Date.now();
 			
-			if(existing_belief.probability <= AGENT_PROBABILITY_THRESHOLD){
+			if (existing_belief.probability <= AGENT_PROBABILITY_THRESHOLD &&
+				!beliefs.allies.find((a) => a.id === existing_belief.id)){
 				beliefs.deleteAgentBelief(existing_belief);
 			}
 		}
@@ -44,16 +39,15 @@ export async function onAgentCallback_simple(agents,beliefs, client) {
 	}
 	
 	if(beliefs.allies.length !== 0 && agents.length > 0){
-		for (const id of beliefs.allies) {
+		for (const a of beliefs.allies) {
 			await (new Say(
-				id,
+				a.id,
 				{
 					topic : "AgentSensing",
 					cnt : agents,
-					token : beliefs.communication_token,
-					msg_id : crypto.randomUUID()
+					token : beliefs.communication_token
 				}
-			)).execute(client,beliefs);
+			)).execute(beliefs);
 		}
 	}
 }
